TerrainGen/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terraingen"
version = "0.1.0"
description = "CLI para geração procedural de terrenos (diamond-square, erosão, rios, lagos)"
readme = "README.md"
authors = [{ name = "TerrainGen Project" }]
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "numpy",
    "pillow",
    "tifffile",
    "whitebox",
    "click>=8.1.0",
    "rich>=13.0.0",
    "rich-click>=1.8.0",
    "gamedev-shared @ file:../Shared",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Multimedia :: Graphics",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.0.0", "ruff>=0.11.0"]

[project.scripts]
terraingen = "terraingen.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
